extract common static viz entrypoint utils
