+++ frontend/src/metabase/static-viz/index.tsx
@@ -1,11 +1,8 @@
 import { setPlatformAPI } from "echarts/core";
 import ReactDOMServer from "react-dom/server";
 
-// eslint-disable-next-line import/order
-import enterpriseOverrides from "ee-overrides";
 import "metabase/utils/dayjs";
 
-import { PLUGIN_CUSTOM_VIZ } from "metabase/plugins/oss/custom-viz";
 import {
   StaticChoropleth,
   getStaticChoroplethSettings,
@@ -13,12 +10,15 @@ import {
 import { StaticVisualization } from "metabase/static-viz/components/StaticVisualization";
 import { LegacyStaticChart } from "metabase/static-viz/containers/LegacyStaticChart";
 import type { LegacyStaticChartType } from "metabase/static-viz/containers/LegacyStaticChart/LegacyStaticChart";
+import {
+  getRawSeriesWithDashcardSettings,
+  initializeContext,
+  toRenderedChart,
+} from "metabase/static-viz/lib/entrypoint";
 import { createStaticRenderingContext } from "metabase/static-viz/lib/rendering-context";
 import { measureTextEChartsAdapter } from "metabase/static-viz/lib/text";
 import { updateStartOfWeek } from "metabase/utils/i18n";
-import MetabaseSettings from "metabase/utils/settings";
 import { extractRemappings, isCartesianChart } from "metabase/visualizations";
-import { extendCardWithDashcardSettings } from "metabase/visualizations/lib/settings/typed-utils";
 import { makeCellBackgroundGetter } from "metabase/visualizations/lib/table_format";
 import { createDataSource } from "metabase/visualizer/utils/data-source";
 import { getVisualizationColumns } from "metabase/visualizer/utils/get-visualization-columns";
@@ -29,13 +29,10 @@ import {
 } from "metabase/visualizer/utils/split-series";
 import type {
   Card,
-  CustomVizPluginId,
-  DashCardVisualizationSettings,
   Dataset,
   DatasetData,
   GeoJSONData,
   RawSeries,
-  SettingKey,
   VisualizerDataSourceId,
   VisualizerVizDefinition,
 } from "metabase-types/api";
@@ -55,6 +52,11 @@ export type {
   RenderedChart,
 } from "./types";
 
+export {
+  initializeContext,
+  registerCustomVizPlugin,
+} from "metabase/static-viz/lib/entrypoint";
+
 setPlatformAPI({
   measureText: measureTextEChartsAdapter,
 });
@@ -66,22 +68,6 @@ function LegacyRenderChart(type: LegacyStaticChartType, options: unknown) {
   return ReactDOMServer.renderToStaticMarkup(
     <LegacyStaticChart type={type} options={options} />,
   );
-}
-
-function getRawSeriesWithDashcardSettings(
-  rawSeries: RawSeries,
-  dashcardSettings: DashCardVisualizationSettings,
-): RawSeries {
-  return rawSeries.map((series, index) => {
-    const isMainCard = index === 0;
-    if (isMainCard) {
-      return {
-        ...series,
-        card: extendCardWithDashcardSettings(series.card, dashcardSettings),
-      };
-    }
-    return series;
-  });
 }
 
 function getVisualizerRawSeries(
@@ -117,30 +103,6 @@ function getVisualizerRawSeries(
       columnValuesMapping,
     },
   ];
-}
-
-export function initializeContext(options: RenderChartOptions) {
-  MetabaseSettings.set("token-features", options.tokenFeatures);
-  MetabaseSettings.set(
-    // Unjustified type cast. FIXME
-    "application-colors" as SettingKey,
-    options.applicationColors,
-  );
-
-  if (typeof enterpriseOverrides === "function") {
-    enterpriseOverrides();
-  }
-
-  MetabaseSettings.set("custom-formatting", options.customFormatting);
-  MetabaseSettings.set("site-locale", options.locale ?? "en");
-}
-
-export function registerCustomVizPlugin(
-  factory: Parameters<typeof PLUGIN_CUSTOM_VIZ.registerCustomVizPlugin>[0],
-  identifier: string,
-  pluginId: CustomVizPluginId,
-) {
-  PLUGIN_CUSTOM_VIZ.registerCustomVizPlugin(factory, identifier, pluginId);
 }
 
 function RenderChart(
@@ -255,7 +217,7 @@ export function renderChart(input: RenderChartInput): RenderedChart {
         input.options,
       );
   }
-  return { type: content.startsWith("<svg") ? "svg" : "html", content };
+  return toRenderedChart(content);
 }
 
 function buildCellBackgroundGetter(
